agrego opcion de buscar producto, corrigo error de importar font, modifico el modal de editar producto para que se vea usando la db.json falsa, corrijo estilos,etc

+++ js/mostrarProductos.js
@@ -4,8 +4,11 @@ const lista = document.querySelector("[data-lista]");
 const editModal = document.getElementById('editModal');
 const spanClose = document.querySelectorAll(".close");
 const formEditarProducto = document.getElementById('formEditarProducto');
+const searchBar = document.getElementById('searchBar');
 let currentEditingProductId = null;
+let productos = [];
 
+//#region CREAR TARJETAS
 // Función para crear una tarjeta de producto
 function crearCard(id, imagen, nombre, precio) {
     const producto = document.createElement("div");
@@ -27,55 +30,76 @@ function crearCard(id, imagen, nombre, precio) {
     `;
     return producto;
 }
-
+//#region LISTAR PRODUCTOS
 // Función para listar los productos
 async function listarProductos() {
     try {
-        const productos = await conexionAPI.listarProductos();
+        productos = await conexionAPI.listarProductos();
         console.log('Datos a procesar:', productos);
-
-        if (productos.length === 0) {
-            const mensaje = document.createElement("p");
-            mensaje.textContent = "No hay productos disponibles.";
-            mensaje.classList.add("no-productos");
-            lista.appendChild(mensaje);
-        } else {
-            productos.forEach(producto => {
-                const card = crearCard(producto.id, producto.imagen, producto.nombre, producto.precio);
-                lista.appendChild(card);
-            });
-            agregarEventListeners();
-        }
+        mostrarProductos(productos);
     } catch (error) {
         console.error("Error al obtener los datos de los productos: ", error);
     }
 }
 
+function mostrarProductos(productos) {
+    lista.innerHTML = '';
+    if (productos.length === 0) {
+        const mensaje = document.createElement("p");
+        mensaje.textContent = "No hay productos disponibles.";
+        mensaje.classList.add("no-productos");
+        lista.appendChild(mensaje);
+    } else {
+        productos.forEach(producto => {
+            const card = crearCard(producto.id, producto.imagen, producto.nombre, producto.precio);
+            lista.appendChild(card);
+        });
+        agregarEventListeners();
+    }
+}
+
+//#region ELIMINAR PRODUCTO
 // Función para agregar el evento al boton de ELIMINAR
 function agregarEventListeners() {
     const botonesEliminar = document.querySelectorAll(".btn-delete");
     botonesEliminar.forEach(boton => {
         boton.addEventListener("click", async (event) => {
             const id = event.target.closest(".btn-delete").dataset.id;
-            try {
-                const response = await fetch(`https://my-json-server.typicode.com/FraNkoRasia/fake-json/productos/${id}`, {
-                    method: 'DELETE',
-                });
 
-                if (response.ok) {
-                    alert('Producto eliminado con éxito');
-                    const productoElement = document.querySelector(`.contenedor-producto[data-id="${id}"]`);
-                    productoElement.remove();
-                } else {
+            // Mostrar la ventana de confirmación
+            const confirmacion = confirm("¿Seguro deseas eliminar el producto?");
+
+            if (confirmacion) {
+                try {
+                    const response = await fetch(`https://my-json-server.typicode.com/FraNkoRasia/fake-json/productos/${id}`, {
+                        method: 'DELETE',
+                    });
+
+                    if (response.ok) {
+                        //Producto eliminado
+                        const productoElement = document.querySelector(`.contenedor-producto[data-id="${id}"]`);
+                        productoElement.remove();
+
+                        // Verificar si quedan productos en la lista
+                        const remainingProducts = document.querySelectorAll(".contenedor-producto");
+                        if (remainingProducts.length === 0) {
+                            const mensaje = document.createElement("p");
+                            mensaje.textContent = "No hay productos disponibles.";
+                            mensaje.classList.add("no-productos");
+                            lista.appendChild(mensaje);
+                        }
+                    } else {
+                        alert('Error al eliminar el producto');
+                    }
+                } catch (error) {
+                    console.error('Error:', error);
                     alert('Error al eliminar el producto');
                 }
-            } catch (error) {
-                console.error('Error:', error);
-                alert('Error al eliminar el producto');
             }
         });
     });
-
+    //#region EDITAR PRODUCTO
+    // Función para agregar el evento al boton de EDITAR
     const botonesEditar = document.querySelectorAll(".btn-edit");
     botonesEditar.forEach(boton => {
         boton.addEventListener("click", (event) => {
@@ -95,8 +119,8 @@ function agregarEventListeners() {
             editModal.style.display = "block";
         });
     });
-
-
+    //#region COMPRAR PRODUCTO
+    // Función para seleccionar un producto y redireccionar a otra vista
     const imagenesProducto = document.querySelectorAll(".contenedor-producto .imagen-producto");
     imagenesProducto.forEach(imagen => {
         imagen.addEventListener("click", (event) => {
@@ -130,6 +154,7 @@ function agregarEventListeners() {
         }
     });
 
+    //#region FUNCION EDITAR
     formEditarProducto.addEventListener("submit", async (event) => {
         event.preventDefault();
 
@@ -156,7 +181,14 @@ function agregarEventListeners() {
 
             if (response.ok) {
                 alert('Producto actualizado con éxito');
-                location.reload(); // Recargar la página para ver los cambios
+
+                // Actualizo el producto en el DOM sin recargar la página
+                const productoElement = document.querySelector(`.contenedor-producto[data-id="${id}"]`);
+                productoElement.querySelector('.nombreProducto').textContent = nombre;
+                productoElement.querySelector('.precioProducto').textContent = `$${precio}`;
+                productoElement.querySelector('.imagen-producto').src = imagen;
+
+                editModal.style.display = "none";
             } else {
                 alert('Error al actualizar el producto');
             }
@@ -167,4 +199,17 @@ function agregarEventListeners() {
     });
 }
 
+// Función de búsqueda
+function buscarProductos() {
+    searchBar.addEventListener('input', () => {
+        const searchTerm = searchBar.value.toLowerCase();
+        const productosFiltrados = productos.filter(producto =>
+            producto.nombre.toLowerCase().includes(searchTerm)
+        );
+        mostrarProductos(productosFiltrados);
+    });
+}
+
+buscarProductos();
 listarProductos();
+
